feat: 커스텀 모바일 헤더 페이지 전용 에러바운더리 생성
